backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
google-generativeai
pydantic
requests